Refactored exception handling
