+++ app/exceptions/exceptions.py
@@ -1,6 +1,12 @@
 from typing import Optional
 
 
+class UserNotFoundException(Exception):
+    def __init__(self, user_uuid: str):
+        self.message = f"User {user_uuid} not found"
+        super().__init__(self.message)
+
+
 class GroupNotFoundException(Exception):
     def __init__(self, group_uuid: Optional[str], custom_message: str = None):
         if custom_message:
