+++ app/main.py
@@ -1,14 +1,9 @@
 from fastapi import FastAPI
-from app.database.database import initialize_db
 from app.api.group import router as group_router
 from app.api.user import router as user_router
 from app.database.database_healt_check import check_db_health
-from app.exceptions.group_not_found_exception import GroupNotFoundException
-from app.exceptions.user_not_found_exception import UserNotFoundException
-from app.exceptions.exception_handlers import (
-    user_not_found_exception_handler,
-    group_not_found_exception_handler,
-)
+
+
 from app.services.startup import lifespan
 
 app = FastAPI(lifespan=lifespan)
@@ -30,9 +25,6 @@ def healthcheck():
         return {"status": "healthy"}
     return {"status": "unhealthy"}
 
-
-app.add_exception_handler(UserNotFoundException, user_not_found_exception_handler)
-app.add_exception_handler(GroupNotFoundException, group_not_found_exception_handler)
 
 app.include_router(user_router)
 app.include_router(group_router)
